Implemented DataManager class for doing CRUD operations on a seperate thread
